Small refactor, fixed issue with unbound entry
- small refactoring for assembler.cpp, added a bunch of commits
- added a unit test to ensure that unbound entries are properly reported
  (they were not)

+++ gdextension/cpp/magix_vm/src/magix_vm/doctest_helper.hpp
@@ -15,7 +15,7 @@ template <class ItA, class SentA, class ItB, class SentB, class Cmp = std::equal
 auto
 doctest_range_eq_impl(
     const char *file,
-    size_t line,
+    int line,
     const doctest::String &name_a,
     const doctest::String &name_b,
     ItA it_a,
@@ -69,7 +69,7 @@ template <class RA, class RB, class Cmp = std::equal_to<>>
 auto
 doctest_range_eq_impl(
     const char *file,
-    size_t line,
+    int line,
     const doctest::String &name_a,
     const doctest::String &name_b,
     const RA &ra,
